move option parsing to separate package
