+++ agent/src/main/java/com/teamscale/jacoco/agent/Agent.java
@@ -5,6 +5,7 @@
 +-------------------------------------------------------------------------*/
 package com.teamscale.jacoco.agent;
 
+import com.teamscale.jacoco.agent.options.AgentOptions;
 import com.teamscale.jacoco.agent.store.IXmlStore;
 import com.teamscale.jacoco.agent.store.UploadStoreException;
 import com.teamscale.jacoco.agent.util.Benchmark;
@@ -41,9 +42,9 @@ public class Agent extends AgentBase {
 	/** Stores the XML files. */
 	protected final IXmlStore store;
 
-	/** Constructor. */
-	/*package*/ Agent(AgentOptions options,
-					  Instrumentation instrumentation) throws IllegalStateException, UploadStoreException {
+	public Agent(AgentOptions options, Instrumentation instrumentation)
+			throws IllegalStateException, UploadStoreException {
+
 		super(options);
 
 		store = options.createStore(instrumentation);
@@ -65,7 +66,8 @@ public class Agent extends AgentBase {
 
 	@Override
 	protected void initServerEndpoints() {
-		get("/partition", (request, response) -> Optional.ofNullable(options.teamscaleServer.partition).orElse(""));
+		get("/partition", (request, response) ->
+				Optional.ofNullable(options.getTeamscaleServerOptions().partition).orElse(""));
 
 		post("/dump", this::handleDump);
 		post("/reset", this::handleReset);
@@ -100,7 +102,7 @@ public class Agent extends AgentBase {
 
 		logger.debug("Changing partition name to " + partition);
 		controller.setSessionId(partition);
-		options.teamscaleServer.partition = partition;
+		options.getTeamscaleServerOptions().partition = partition;
 
 		response.status(204);
 		return "";
