+++ agent/src/test/java/com/teamscale/jacoco/agent/options/AgentOptionsParserTest.java
@@ -1,5 +1,6 @@
-package com.teamscale.jacoco.agent;
+package com.teamscale.jacoco.agent.options;
 
+import com.teamscale.jacoco.agent.options.AgentOptionsParser;
 import com.teamscale.report.util.CommandLineLogger;
 import org.junit.Test;
